[refactor] : header 컴포넌트 수정
1. header 컴포넌트가 재사용이 많이 되기 때문에 components 디렉토리 바로 밑에 Header.jsx를 두었습니다.
